% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R, R/track_generic.R
\name{assign_circos_track_default}
\alias{assign_circos_track_default}
\title{Assign default values to a circos track object}
\usage{
assign_circos_track_default(
  track,
  default_color = "grey20",
  default_point_size = 0.5,
  default_ylim = NULL,
  default_bg_color = "white",
  default_bg_highlight_color = "grey80"
)

assign_circos_track_default(
  track,
  default_color = "grey20",
  default_point_size = 0.5,
  default_ylim = NULL,
  default_bg_color = "white",
  default_bg_highlight_color = "grey80"
)
}
\arguments{
\item{track}{A track object generated with the \code{\link{circos_track}} function)}

\item{default_color}{Default color for a track when not specified in track data (default: "grey20")}

\item{default_point_size}{Default point size for a track of type "points" when not specified in track data (default: 0.5)}

\item{default_ylim}{Default y-axis limits for a track when not specified in track data (default: NULL, i.e. infer from data)}

\item{default_bg_color}{Default background color for standard sectors in the track (default: "white")}

\item{default_bg_highlight_color}{Default background color for highlighted sectors in the track (default: "grey80")}
}
\value{
A track object with default values for properties not specified by the user

A track object with default values for properties not specified by the user
}
\description{
Assign default values to all properties for which the value was not specified by the user (e.g. value is NULL)

Assign default values to all properties for which the value was not specified by the user (e.g. value is NULL)
}
\examples{
track_data <- assign_circos_track_default(track_data, default.alpha = 0.75)

track_data <- assign_circos_track_default(track_data, default.alpha = 0.75)

}
