% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{assign_defaults}
\alias{assign_defaults}
\title{Assign default values to a track object}
\usage{
assign_defaults(track, defaults)
}
\arguments{
\item{track}{A track object generated with the \code{\link{track}} function)}

\item{defaults}{A named list with default values for all properties.}
}
\value{
A track object with values for all properties.
}
\description{
Assign default values to all properties for which the value
was not specified by the user (e.g. value is NA).
}
\examples{
track <- track("Fst", label = "Window FST")
defaults <- list(colors = "red", alpha = 0.5)
track <- assign_defaults(track, defaults)
}
