% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{assign_manhattan_background_colors}
\alias{assign_manhattan_background_colors}
\title{Assign color to manhattan background data}
\usage{
assign_manhattan_background_colors(data, track)
}
\arguments{
\item{data}{Data frame with contig, start, and end for each background rectangle}

\item{track}{Track object for the current plot, generated with the \code{\link{manhattan_track}} function}
}
\value{
A data frame with columns:
contig | start | end | color
}
\description{
Assign color to each background rectangle from a backgrounds data frame for the \code{\link{plot_track_manhattan}} function
}
\examples{
track <- manhattan_track("Fst")
background_data <- assign_manhattan_background_colors(background_data, track)

}
