% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{assign_manhattan_track_default}
\alias{assign_manhattan_track_default}
\title{Assign default values to a manhattan track object}
\usage{
assign_manhattan_track_default(
  track,
  default.point.color = c("dodgerblue3", "darkgoldenrod2"),
  default.bg.color = c("white", "grey85"),
  default.point.size = 0.5,
  default.ylim = NULL
)
}
\arguments{
\item{track}{A track object generated with the \code{\link{manhattan_track}} function)}

\item{default.point.color}{Default point color when not specified in track data (default: c("dodgerblue3", "darkgoldenrod2"))}

\item{default.bg.color}{Default background color when not specified in track data (default: c("white", "grey85"))}

\item{default.point.size}{Default point size for a track when not specified in track data (default: 0.5)}

\item{default.ylim}{Default y-axis limits for a track when not specified in track data (default: NULL, i.e. infer from data)}
}
\value{
A track object with default values for properties not specified by the user
}
\description{
Assign default values to all properties for which the value was not specified by the user (i.e. value is NULL)
}
\examples{
track_data <- assign_manhattan_track_default(track_data, default.point.size = 0.75)

}
