% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{assign_region_track_default}
\alias{assign_region_track_default}
\title{Assign default values to a region track object}
\usage{
assign_region_track_default(
  track,
  default.color = "grey20",
  default.alpha = 1,
  default.type = "ribbon",
  default.point.size = 0.5,
  default.ylim = NULL,
  default.major.lines.y = TRUE,
  default.major.lines.x = FALSE,
  default.legend.position = "right"
)
}
\arguments{
\item{track}{A track object generated with the \code{\link{region_track}} function)}

\item{default.color}{Default color for a track when not specified in track data (default: "grey20")}

\item{default.alpha}{Default alpha value for a track when not specified in track data (default: 1)}

\item{default.type}{Default plot type for a track when not specified in track data (default: "ribbon")}

\item{default.point.size}{Default point size for a track of type "points" when not specified in track data (default: 0.5)}

\item{default.ylim}{Default y-axis limits for a track when not specified in track data (default: NULL, i.e. infer from data)}

\item{default.major.lines.y}{If TRUE, reference lines will be plotted for the y axis if not specified in track data (default: TRUE)}

\item{default.major.lines.x}{If TRUE, reference lines will be plotted for the x axis if not specified in track data (default: FALSE)}

\item{default.legend.position}{Default legend position when not specified in track data (default: "none")}
}
\value{
A track object with default values for properties not specified by the user
}
\description{
Assign default values to all properties for which the value was not specified by the user (e.g. value is NULL)
}
\examples{
track_data <- assign_region_track_default(track_data, default.alpha = 0.75)

}
