% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{assign_values}
\alias{assign_values}
\title{Assign value to track property}
\usage{
assign_values(metrics, property, values)
}
\arguments{
\item{metrics}{Named list of metrics properties/}

\item{colors}{Colors, can be a string (e.g. "grey20") and will be
applied to all metrics, a vector of size equal to the number of metrics
(e.g. c("red", "blue") for two metrics), or a color scale object <TBD>.}

\item{ylim}{Y-axis limits, a vector with values c(min, max).}

\item{point_size}{Point size for plots of type "points". Values can be a
float (e.g. 0.5) and will be applied to all metrics, or a vector of
size equal to the number of metrics (e.g. c(1, 1.5, 3) for three metrics).}

\item{alpha}{Alpha value. Values can be a float (e.g. 0.5) and will
be applied to all metrics, or a vector of size equal to the number of metrics
(e.g. c(1, 0.5, 0.75) for three metrics).}

\item{type}{Plot type for circos and region plots, either "ribbon" or
"points". Values can be a string (e.g. "ribbon") and will be applied to all
metrics, or a vector of size equal to the number of metrics
(e.g. c("points", "ribbon") for two metrics). Not used in manhattan plots.}
}
\value{
The correct value to assign to the property.
}
\description{
Assign value to a single track property, handling single vs
multi metrics value definitions.
}
\examples{
track <- single_metric_track("Fst")
assign_value(track$metrics, "alpha", 0.6)
}
