% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R, R/track_generic.R
\name{circos_track}
\alias{circos_track}
\title{Create circos track object}
\usage{
circos_track(
  metrics,
  label = NULL,
  color = NULL,
  point.size = NULL,
  ylim = NULL,
  bg.color = NULL,
  bg.highlight.color = NULL
)

circos_track(
  metrics,
  label = NULL,
  color = NULL,
  point.size = NULL,
  ylim = NULL,
  bg.color = NULL,
  bg.highlight.color = NULL
)
}
\arguments{
\item{metrics}{Metrics included in the track. Metrics should correspond to column names in the data frame used as input
data in the plot (output of the \code{\link{load_genome_input}} function). Possible values: a string if the track includes
a single metric (e.g. "Fst"), or a vector if the track includes several metrics (e.g. c("Snps_females", "Snps_males"))}

\item{label}{Track label, NULL to set the label to the metric name for single-metric tracks. Label has to be specified for multi-metrics tracks (default: NULL)}

\item{color}{Track color. Values can be a string (e.g. "grey20") and will then be applied to all metrics, or a vector of size equal to the number of metrics
(e.g. c("red", "blue") for two metrics)}

\item{point.size}{Point size for plots of type "points". Values can be a float (e.g. 0.5) and will then be applied to all metrics,
or a vector of size equal to the number of metrics (e.g. c(1, 1.5, 3) for three metrics)}

\item{ylim}{Vector of y-axis limits for the track; if NULL, infer directly from data}

\item{bg.color}{Background color for standard sectors in the track (single value or vector of length equal to the number of sectors in the plot)}

\item{bg.highlight.color}{Background color for highlighted sectors in the track (single value)}
}
\value{
A named list with the value of each track property

A named list with the value of each track property
}
\description{
Generate an object storing all properties for a circos track

Generate an object storing all properties for a circos track
}
\examples{
# Single metric
track_data <- circos_track("Fst", color = "grey70", point.size = 0.75)

# Multiple metrics
track_data <- circos_track(c("Snp_females", "Snp_males"), color = c("firebrick2", "dodgerblue3"))

# Single metric
track_data <- circos_track("Fst", color = "grey70", point.size = 0.75)

# Multiple metrics
track_data <- circos_track(c("Snp_females", "Snp_males"), color = c("firebrick2", "dodgerblue3"))

}
