% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsex_marker_depths.R
\name{cluster_marker_depths}
\alias{cluster_marker_depths}
\title{Cluster marker depths data}
\usage{
cluster_marker_depths(
  data,
  min_depth = NA,
  max_depth = 150,
  distance_method = "euclidean",
  clustering_method = "ward.D"
)
}
\arguments{
\item{data}{A table of marker depths obtained with the
\code{\link{load_table}} function.}

\item{min_depth}{Minimum depth to consider a marker present in an individual:
depths lower than this value will be set to 0 (default: NA, i.e. get value
from input data).}

\item{max_depth}{Maximum marker depth allowed in an individual: depths higher
than this value will be set to this value (default: 100).}

\item{distance_method}{Method to use to compute the distance matrix, directly
passed to \code{\link[stats]{dist}}. Possible values: "euclidean", "maximum",
"manhattan", "canberra", "binary", and "minkowski" (default: "euclidean")}

\item{clustering_method}{Method to use for clustering, directly passed to
\code{\link[stats]{hclust}}. Possible values: "ward.D", "ward.D2", "single",
"complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC),
and "centroid" (= UPGMC) (default: "ward.D").}
}
\value{
A list with the following elements:
\item{data}{A data frame of marker depths with individuals and markers
ordered based on the clustering results}
\item{individuals}{Individuals clustering results}
\item{markers}{Markers clustering results}
\item{distribution}{Distribution of depth values}
}
\description{
Cluster individuals and markers by depth values in a table of
marker depths.
}
\examples{
clustering_data <- cluster_marker_depths(data,
                                         min_depth = 1,
                                         max_depth = 100,
                                         distance_method = "binary",
                                         clustering_method = "complete")
}
