% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{configure_backgrounds}
\alias{configure_backgrounds}
\title{Configure sector background colors}
\usage{
configure_backgrounds(track, highlight, highlight_bg_color)
}
\arguments{
\item{highlight}{Vector containing the names or identifiers of contigs or
chromosomes to highlight in the circos plot.}

\item{highlight_bg_color}{Background color for highlighted sectors.}

\item{tracks}{List of track data objects for each track to include in the
plot. Track data objects are generated with the \code{\link{track}},
\code{\link{single_metric_track}}, or \code{\link{multi_metrics_track}}
functions.}
}
\value{

}
\description{
Configure background colors for all sectors in circos plot
}
\examples{
track <- single_metric_track("Fst", colors = "grey70", ylim = c(0, 1))
highlight <- c("LG07")
track <- configure_backgrounds(track, highlight, "grey70")
}
