% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{configure_track}
\alias{configure_track}
\title{Configure track data from default values}
\usage{
configure_track(track, defaults, data, region = NA)
}
\arguments{
\item{track}{A track object generated with the \code{\link{track}} function).}

\item{defaults}{A named list with default values for all properties.}

\item{data}{Genomic data (e.g. result of PSASS or RADSex loaded
with the \code{\link{load_genome_metrics}} function).}

\item{region}{Region to plot, output of the \code{\link{parse_region}}
function, or NA to retain the entire genome (default: NA)}
}
\value{
A track object with values for all properties and track data stored
in track$data.
}
\description{
Wrapper around several track value assignment functions to
assign default values and generate track data.
}
\examples{
track <- track("SNP_males", label = "Male-specific SNPs")
defaults <- list(colors = "red", alpha = 0.5)
data <- load_genome_metrics("window.tsv")
track <- configure_track(track, defaults, data)
}
