% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_mb}
\alias{convert_to_mb}
\title{Convert position to Mb}
\usage{
convert_to_mb(x, n = 0)
}
\arguments{
\item{x}{Genomic position to convert, integer}

\item{n}{Number of digits to round the value, integer (default: 0)}
}
\value{
The converted value, floating point number
}
\description{
Convert a genomic position to Mega bp (round to n digits).
This function can be called directly to generate labels in plot scales.
e.g. ggplot2::scale_y_continuous(labels = convert_to_mb)
}
\examples{
x_mb <- convert_to_mb(1267356, 2)
x_mb
> 1.27
}
