% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R, R/track_generic.R
\name{create_circos_track_data}
\alias{create_circos_track_data}
\title{Create circos track data}
\usage{
create_circos_track_data(data, track)

create_circos_track_data(data, track)
}
\arguments{
\item{data}{Genomic data (e.g. result of PSASS or RADSex loaded with the \code{\link{load_genome_input}} function)}

\item{track}{Track object for the current plot, generated with the \code{\link{circos_track}} function}
}
\value{
A data frame with columns:
Contig_plot | Position_plot | Metric 1 | Metric 1 colors | ... | Metric N | Metric N colors

A data frame with columns:
Contig_plot | Position_plot | Metric 1 | Metric 1 colors | ... | Metric N | Metric N colors
}
\description{
Create input data frame for the \code{\link{plot_track_circos}} function from the genomic data and the track information

Create input data frame for the \code{\link{plot_track_circos}} function from the genomic data and the track information
}
\examples{
genomic_data <- load_genome_input("psass_window.tsv")
track <- circos_track("Fst")

region_data <- create_circos_track_data(genomic_data, track)

genomic_data <- load_genome_input("psass_window.tsv")
track <- circos_track("Fst")

region_data <- create_circos_track_data(genomic_data, track)

}
