% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsex_map.R
\name{create_palette}
\alias{create_palette}
\title{Create bias color palette}
\usage{
create_palette(bias_colors)
}
\arguments{
\item{bias_colors}{A vector of length 3 defining colors for the extremes in
the bias track, i.e. -1, 0, and 1.}
}
\value{
A function returning the color associated to a bias value.
}
\description{
Generate a function to assign a color to a bias value based on
colors defined in bias_colors.
}
\examples{
palette <- create_palette(c("blue", "black", "yellow"))
palette(0.25)
}
