% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{create_region_track_data}
\alias{create_region_track_data}
\title{Create region track data}
\usage{
create_region_track_data(data, region.info, track)
}
\arguments{
\item{data}{Genomic data (e.g. result of PSASS or RADSex loaded with the \code{\link{load_genome_input}} function)}

\item{region.info}{Information on the region to plot, output of the \code{\link{parse_region}} function}

\item{track}{Track object for the current plot, generated with the \code{\link{region_track}} function}
}
\value{
A data frame with columns:
Position | Metric 1 | Metric 1 colors | ... | Metric N | Metric N colors
}
\description{
Create input data frame for the \code{\link{plot_track_region}} function from the genomic data and the track information
}
\examples{
genomic_data <- load_genome_input("psass_window.tsv")
region_info <- parse_region("Chr01:0-1500000")
track <- region_track("Fst")

region_data <- create_region_track_data(genomic_data, region_info, track)

}
