% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{create_track_data}
\alias{create_track_data}
\title{Create track data}
\usage{
create_track_data(track, data, region = NA)
}
\arguments{
\item{track}{Track object for the current plot, generated with the
\code{\link{track}} function.}

\item{data}{Genomic data (e.g. result of PSASS or RADSex loaded
with the \code{\link{load_genome_metrics}} function).}

\item{region}{Region to plot, output of the \code{\link{parse_region}}
function, or NA to retain the entire genome (default: NA)}
}
\value{
A data frame with columns:
Contig | Position | Metric 1 | Metric 1 colors |  Metric N | Metric N colors
}
\description{
Create an input data frame for the track plotting functions
}
\examples{
genomic_data <- load_genome_metrics("psass_window.tsv")
track <- track("Fst")
track <- create_track_data(genomic_data, track)
}
