% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_genome_metrics.R
\name{detect_chromosomes}
\alias{detect_chromosomes}
\title{Detect chromosomes}
\usage{
detect_chromosomes(data)
}
\arguments{
\item{data}{Genome metrics data frame: data frame with columns
Contig | Position | Length | <Metric1> | <Metric2> ...}
}
\value{
A named vector with contig identifiers as names and
chromosome names as values
}
\description{
Automatically detect chromosomes in a genome metrics data frame.
Contigs are identified as chromosomes if their name starts with
"CHR", "LG", or "NC" (case unsensitive).
}
\examples{
chromosome_names <- detect_chromosomes(data)
}
