% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw_barplot}
\alias{draw_barplot}
\title{Draw a barplot for a metric}
\usage{
draw_barplot(
  data,
  group_labels = NA,
  group_colors = c("dodgerblue3", "firebrick2"),
  alpha = 0.8,
  output_file = NA,
  width = 8,
  height = 8,
  res = 300,
  title = NA,
  ylab = NA,
  major_lines_y = TRUE,
  pretty_y_scale = TRUE,
  order_by = "Metric"
)
}
\arguments{
\item{data}{A data frame with columns "Sample | Group | <Metric>". The
metric is assumed to have continuous values.}

\item{group_labels}{Vector specifying the label to associate to each group
in the plot legend, or NA to use the group names (default: NA).}

\item{group_colors}{Vector specyfing the color to use for each group in the
plot (default: c("dodgerblue3", "firebrick2")).}

\item{alpha}{Alpha value for the bars, a float (default: 0.8).}

\item{output_file}{Path to an output file for the generated barplot,
or NA to plot in the current R device (default: NA).}

\item{width}{Plot width when plotting to an output file, in inches
(default: 12).}

\item{height}{Plot height when plotting to an output file, in inches
(default: 4).}

\item{res}{Image resolution when plotting to an output file, in dpi
(default: 300).}

\item{title}{Plot title, or NA for no title (default: NA).}

\item{ylab}{Title of the y axis, or NA to use the metric name (default: NA).}

\item{major_lines_y}{If TRUE, reference lines will be plotted for the y axis,
equivalent to panel.grid.major.y in \code{\link{ggplot2::theme}}
(default: TRUE).}

\item{pretty_y_scale}{If TRUE, generate pretty labels for the y-axis scale
and add the unit (M. or K.) to the axis title (default: TRUE).}

\item{order_by}{Value to use to order bars in the plot, either "Group",
"Sample", or "Metric" (default: "Metric").}
}
\value{
A ggplot object for the plot
}
\description{
Draw a barplot of the value of a metric for each Sample
in a data frame with columns: "Sample | Group | <Metric>".
}
\examples{
data <- load_table("depth.tsv")
data <- data[, c("Sample", "Group", "Markers")]
markers_plot <- draw_barplot(data,
                             group_labels = c("Males", "Females"),
                             group_colors = c("blue", "red"),
                             ylab = "Number of reads")
}
