% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{draw_circos_track}
\alias{draw_circos_track}
\title{Plot a circos track}
\usage{
draw_circos_track(
  track,
  top_track = FALSE,
  sector_titles_expand = 1.3,
  first_sector = NA
)
}
\arguments{
\item{track}{Track object storing properties for the current track,
generated with the \code{\link{track}}, \code{\link{single_metric_track}},
or \code{\link{multi_metrics_track}} functions.}

\item{top_track}{If TRUE, x-axis labels and sector names will be added to
the track (default: FALSE).}

\item{sector_titles_expand}{Manually set the space between sector titles and
x-axis as a multiple of ymax (default: 1.3).}

\item{first_sector}{Name of the first sector to draw y-axis on
(default: NA).}
}
\description{
Plot a single track for a circos plot
}
\examples{
data <- load_genome_metrics("window.tsv")
defaults <- list(colors = "blue", alpha = 0.5)
track <- single_metric_track("SNP_males", label = "Male-specific SNPs")
track <- configure_track(track, defaults, data)
circlize::circos.par("track.height" = 12)
draw_circos_track(track, top_track=TRUE)
}
