% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw_density}
\alias{draw_density}
\title{Draw density for a metric}
\usage{
draw_density(
  data,
  color = "grey60",
  alpha = 0.8,
  output_file = NA,
  width = 12,
  height = 6,
  res = 300,
  title = NA,
  ylab = NA,
  xlab = NA,
  major_lines_y = TRUE
)
}
\arguments{
\item{data}{A vector of values.}

\item{color}{Color of the plot (default: "grey60").}

\item{alpha}{Alpha value for the bars, a float (default: 0.8).}

\item{output_file}{Path to an output file for the generated barplot,
or NA to plot in the current R device (default: NA).}

\item{width}{Plot width when plotting to an output file, in inches
(default: 12).}

\item{height}{Plot height when plotting to an output file, in inches
(default: 4).}

\item{res}{Image resolution when plotting to an output file, in dpi
(default: 300).}

\item{title}{Plot title, or NA for no title (default: NA).}

\item{ylab}{Title of the y axis, or NA to use the metric name (default: NA).}

\item{xlab}{Title of the x axis, or NA for no title (default: NA).}

\item{major_lines_y}{If TRUE, reference lines will be plotted for the y axis,
equivalent to panel.grid.major.y in \code{\link{ggplot2::theme}}
(default: TRUE).}

\item{pretty_y_scale}{If TRUE, generate pretty labels for the y-axis scale
and add the unit (M. or K.) to the axis title (default: TRUE).}
}
\value{
A ggplot object for the plot
}
\description{
Draw a density plot for a vector of values.
}
\examples{
data <- load_table("freq.tsv")
data <- rep(data$Frequency, data$Count)
freq_plot <- draw_density(data)
}
