% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{draw_manhattan_plot}
\alias{draw_manhattan_plot}
\title{Draw manhattan plot}
\usage{
draw_manhattan_plot(
  data,
  contig.lengths,
  tracks,
  output.file = NULL,
  width = 14,
  track.height = 6,
  res = 300,
  x.title = "Chromosomes",
  show.chromosomes.names = TRUE,
  chromosomes.as.numbers = FALSE,
  default.point.color = c("dodgerblue3", "darkgoldenrod2"),
  default.bg.color = c("grey85", "white"),
  default.point.size = 0.5,
  default.ylim = NULL
)
}
\arguments{
\item{data}{Genomic data (e.g. result of PSASS or RADSex loaded with the \code{\link{load_genome_input}} function)}

\item{contig.lengths}{Contig lengths from the output of the \code{\link{load_genome_input}} function}

\item{tracks}{List of tracks to plot. Tracks can be generated with the \code{\link{manhattan_track}} function}

\item{output.file}{Path to an output file for the generated Manhattan plot, or NULL to plot in the current R device (default: NULL)}

\item{width}{Plot width when plotting to an output file, in inches (default: 12)}

\item{track.height}{Height of a single track when plotting to an output file, in inches (default: 6)}

\item{res}{Image resolution when plotting to an output file, in dpi (default: 300)}

\item{x.title}{Title of the x-axis}

\item{show.chromosomes.names}{If TRUE, display chromosome names on the x axis}

\item{chromosomes.as.numbers}{If TRUE, display chromosome numbers instead of names for readability}

\item{default.point.color}{Default color for a track when not specified in track data (default: "grey20")}

\item{default.bg.color}{Default background color for a track when not specified in the track (default: "white")}

\item{default.point.size}{Default point size for a track when not specified in track data (default: 0.25)}

\item{default.ylim}{Default y-axis limits for a track when not specified in track data (default: NULL, i.e. infer from data)}
}
\description{
Generate a Manhattan plot with multiple track for the entire genome
}
\examples{
a
}
