% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{draw_manhattan_track}
\alias{draw_manhattan_track}
\title{Plot a manhattan track}
\usage{
draw_manhattan_track(
  track,
  bottom_track = FALSE,
  x_title = "Chromosomes",
  show_chromosomes_names = TRUE,
  chromosomes_as_numbers = FALSE
)
}
\arguments{
\item{track}{Track object storing properties for the current track,
generated with the \code{\link{track}} function.}

\item{bottom_track}{If TRUE, x-axis labels and title will be added to
the plot (default: FALSE).}

\item{x_title}{Title of the x-axis (default: "Chromosomes")}

\item{show_chromosomes_names}{If TRUE, display chromosome names on the x axis
(default: TRUE)}

\item{chromosomes_as_numbers}{If TRUE, display chromosome numbers instead of
names for readability (default: FALSE)}
}
\value{
A ggplot object for the track
}
\description{
Plot a single track for a manhattan plot
}
\examples{
data <- load_genome_metrics("window.tsv")
defaults <- list(colors = c("yellow", "blue"))
track <- track("SNP_males", label = "Male-specific SNPs")
track <- configure_track(track, defaults, data)
track_plot <- draw_manhattan_track(track, bottom_track=TRUE)
}
