% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{draw_region_track}
\alias{draw_region_track}
\title{Plot a region track}
\usage{
draw_region_track(track, region, bottom_track = FALSE)
}
\arguments{
\item{track}{Track object storing properties for the current track,
generated with the \code{\link{track}} function.}

\item{region}{Region to plot, output of the \code{\link{parse_region}}
function, or NA to retain the entire genome.}

\item{bottom_track}{If TRUE, x-axis labels and title will be added to
the plot (default: FALSE).}
}
\value{
A ggplot object for the plot
}
\description{
Plot a single track for a genomic region
}
\examples{
data <- load_genome_metrics("psass_window.tsv")
defaults <- list(colors = "blue", alpha = 0.5)
track <- single_metric_track("SNP_males", label = "Male-specific SNPs")
track <- configure_track(track, defaults, data)
draw_region_track(track, "Chr01", bottom_track=TRUE)
}
