% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_columns}
\alias{drop_columns}
\title{Drop columns from a data frame}
\usage{
drop_columns(data, columns)
}
\arguments{
\item{data}{A data frame.}

\item{columns}{A string (i.e. "value") or vector of strings (
i.e. c("X1", "X2")) of columns to drop from the data frame.}
}
\value{
A data frame in which the specified columns were removed.
}
\description{
Remove columns from a data frame by name. Generate a warning
if columns to remove do not exist in the data frame.
}
\examples{
data <- data.frame(x = c(1), y = c(2), z = c(3))
data <- drop_columns(data, c("x", "z"))
}
