% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsex_distrib.R
\name{generate_color_palette}
\alias{generate_color_palette}
\title{Generate a color palette for "distrib" plot}
\usage{
generate_color_palette(bins, colors)
}
\arguments{
\item{bins}{A vector of values to use as bins in the color palette.}

\item{colors}{A vector of two colors used to create the color palette
gradient.}
}
\value{
A color palette vector.
}
\description{
Generate a color palette from the specified bins and colors to
use in radsex "distrib" results plot.
}
