% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_y_scale}
\alias{generate_y_scale}
\title{Generate y-axis scale}
\usage{
generate_y_scale(values, metric, n_ticks = 5, min = NA, max = NA)
}
\arguments{
\item{values}{A vector of values that will be plotted on the y axis.}

\item{metric}{The metric name.}

\item{n_ticks}{Number of ticks to include in the scale (default: 5)}

\item{min}{Minimum value for the scale, or NA to infer from the data
(default: NA)}

\item{max}{Maximum value for the scale, or NA to infer from the data
(default: NA)}
}
\value{
A ggplot2::scale_y_continuous object encoding the scale.
}
\description{
Generate an x-axis scale with a given number of ticks,
automatically converting values to the best unit (M., K., ...).
}
\examples{
nice_y_scale <- generate_y_scale(seq(1000000, 10000000, 1000000),
                                 metric = "Reads", n_ticks = 4)
}
