% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_genome_metrics.R
\name{get_contig_lengths_from_data}
\alias{get_contig_lengths_from_data}
\title{Get contig lengths from genome metrics data}
\usage{
get_contig_lengths_from_data(data, chromosomes = NULL, sortby = "Length")
}
\arguments{
\item{data}{Genome metrics data frame: data frame with columns
Contig | Position | Length | <Metric1> | <Metric2> ...}

\item{chromosomes}{Named vector of chromosome names from
\code{\link{load_chromosome_names}}. If not NULL, only returns lengths for
chromosomes (default: NULL)}

\item{sortby}{Column to sort the data frame by, either "Length" or "Contig"
(default: "Length")}
}
\value{
A data frame with columns <Contig ID> | <Contig Length> sorted by
contig length
}
\description{
Extract contig identifier and contig length columns from a
genome metrics data frame and return a sorted data frame.
}
\examples{
contig_lengths <- get_contig_lengths_from_data(data)
}
