% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{h_line}
\alias{h_line}
\title{Create a h_line object}
\usage{
h_line(
  y,
  label = NA,
  label_x = 0.1,
  label_font_size = 7,
  color = "grey60",
  type = 1,
  size = 1
)
}
\arguments{
\item{y}{Y-axis coordinate for the line.}

\item{label}{Label to add to the line, or NA for no label (default: NA).}

\item{label_x}{Normalized x coordinate for the label (default: 0.1).}

\item{label_font_size}{Font size for label (default: 7).}

\item{color}{Line color (default: "grey60").}

\item{type}{Line type as defined in R "lty" parameter (default: 1).}

\item{size}{Line thickness (default: 1).}
}
\value{
A named list with the value of each line property
}
\description{
Generate an object storing all properties for a horizontal
line
}
\examples{
signif_line <- h_line("p>0.05", color = "red")
}
