% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_genome_metrics.R
\name{load_genome_metrics}
\alias{load_genome_metrics}
\title{Load a genome metrics data file}
\usage{
load_genome_metrics(
  input_file_path,
  chromosomes = NULL,
  detect_chromosomes = TRUE,
  unplaced_label = "Unplaced"
)
}
\arguments{
\item{input_file_path}{Path to a genome metrics file (format described above)}

\item{chromosomes}{Named vector of chromosome names from
\code{\link{load_chromosome_names}} (default: NULL)}

\item{detect_chromosomes}{If TRUE, will consider contigs starting with
"LG", "CHR", or "NC" as chromosomes if no chromosomes were specified
(default: TRUE)}

\item{unplaced_label}{Label for unplaced contigs superscaffold
(default: "Unplaced")}
}
\value{
A list with two elements: "data" = parsed genome metrics data,
"lengths" = lengths of contigs included in the plot
}
\description{
Loads a file containing metric values along a genome.
Format: Contig | Position | Length | <Metric1> | <Metric2> ... with
Contig = contig identifier, Position = position on the contig, Length =
length of the contig, <MetricN> = value of metric N (e.g. Fst) at the
corresponding position on the corresponding contig.
}
\examples{
chromosomes <- load_chromosome_names("chromosomes_names.tsv")
genome_data <- load_genome_metrics("psass_window.tsv",
                                   chromosomes = chromosomes)
}
