% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsex_data.R
\name{load_group_info}
\alias{load_group_info}
\title{Load group info}
\usage{
load_group_info(input_file, groups = NA, group_labels = NA)
}
\arguments{
\item{input_file}{Path to a tabulated file with columns "Individual | Group"}

\item{groups}{Vector of groups to include in the plots, in order; group names
should match column headers in the input data. If NA, groups will be infered
from the input data (default: NA).}

\item{group_labels}{Vector of length 2 with the labels to associate to each
group in the plot axis titles, or NA to use the group names (default: NA).}
}
\value{
A list with the following elements:
\item{individual_groups}{Vector with individuals as names and groups as values}
\item{groups}{Vector of group names}
\item{group_labels}{Vector with groups as names and group labels as values}
\item{counts}{Vector with groups as names and individual counts as values}
}
\description{
Load group information from a popmap file.
}
\examples{
popmap <- load_group_info("popmap.tsv",
                          groups = c("M", "F"),
                          group_labels = c("Males", "Females"))
}
