% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsex_data.R
\name{load_marker_depths}
\alias{load_marker_depths}
\title{Load a standard table file}
\usage{
load_marker_depths(input_file)
}
\arguments{
\item{input_file}{Path to a tabulated file}
}
\value{
A data frame containing the tabulated file data.
}
\description{
Loads a tabulated file with readr
}
\examples{
data <- load_table("subset.tsv")
}
