% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsex_data.R
\name{load_marker_distribution}
\alias{load_marker_distribution}
\title{Load a marker distribution table}
\usage{
load_marker_distribution(input_file, groups = NA, group_labels = NA)
}
\arguments{
\item{input_file}{Path to a table of distribution of markers between groups
obtained with radsex "distrib".}

\item{groups}{Vector of groups to include in the plots, in order; group names
should match column headers in the input data. If NA, groups will be infered
from the input data (default: NA).}

\item{group_labels}{Vector of length 2 with the labels to associate to each
group in the plot axis titles, or NA to use the group names (default: NA).}
}
\value{
A list with the following elements:
\item{data}{A data frame with the distribution of markers between groups}
\item{groups}{Vector of group names}
\item{group_labels}{Vector with groups as names and group labels as values}
\item{counts}{Vector with groups as names and individual counts as values}
}
\description{
Load a distribution of markers between groups obtained with
radsex "distrib"
}
\examples{
data <- load_marker_distribution("distrib.tsv",
                                 groups = c("Pop1", "Pop2"),
                                 group_labels = c("France", "Spain"))
}
