% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genome_metrics.R
\name{load_table}
\alias{load_table}
\title{Load a standard table file}
\usage{
load_table(
  input_file,
  get_properties = TRUE,
  comment_char = "#",
  comment_sep = ";",
  comment_internal_sep = ":"
)
}
\arguments{
\item{input_file}{Path to a tabulated file}

\item{get_properties}{If TRUE, will parse the file to identify comment lines
defining properties. Comment lines have the format:
#' <comment_char>property<comment_internal_sep>value<comment_sep>property...
With default values: #property1:value1;property2:value2...}

\item{comment_char}{Character indicating a comment line (default: "#").}

\item{comment_sep}{Character separating two fields in a comment line
(default: ";").}

\item{comment_internal_sep}{Character separating property and value in a
field from a comment line (default: ":").}
}
\value{
A data frame containing the tabulated file data.
}
\description{
Loads a tabulated file with readr
}
\examples{
data <- load_table("subset.tsv")
}
