% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{manhattan_bg_colors}
\alias{manhattan_bg_colors}
\title{Assign color to manhattan background data}
\usage{
manhattan_bg_colors(track, data)
}
\arguments{
\item{track}{Track object for the current plot, generated with the
\code{\link{track}} function.}

\item{data}{Data frame with contig, start, and end for each
background rectangle.}
}
\value{
A track object containing a backgrounds data frame with columns:
contig | start | end | color.
}
\description{
Assign color to each background rectangle from a backgrounds
data frame for the \code{\link{plot_track_manhattan}} function.
}
\examples{
track <- single_metric_track("Fst")
background_data <- data.frame(contig = c("Chr1", "Chr2"),
                              start = c(0, 1000)
                              end = c(999, 2635))
background_data <- manhattan_bg_colors(track, background_data)
}
