% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{manhattan_track}
\alias{manhattan_track}
\title{Create manhattan track object}
\usage{
manhattan_track(
  metric,
  label = NULL,
  point.color = NULL,
  bg.color = NULL,
  point.size = NULL,
  ylim = NULL
)
}
\arguments{
\item{metric}{Metric represented by the track, which should correspond to a column name in the data frame used as input
data in the plot (output of the \code{\link{load_genome_input}} function)}

\item{label}{Track label, NULL to set the label to the metric name (default: NULL)}

\item{point.color}{Point color, either a string (e.g. "grey20") or a vector of alternating colors (e.g. c("red", "blue") for two colors)}

\item{bg.color}{Background color, either a string (e.g. "white") or a vector of alternating colors (e.g. c("grey85", "white") for two colors)}

\item{point.size}{Point size, directly passed to ggplot}

\item{ylim}{Vector of y-axis limits for the track; if NULL, infer directly from data}
}
\value{
A named list with the value of each track property
}
\description{
Generate an object storing all properties for a manhattan track
}
\examples{
# Single metric
track_data <- manhattan_track("Fst", point.color = c("blue", "yellow"), bg.color = c("white", "grey50"), point.size = 0.75)
}
