% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{metric}
\alias{metric}
\title{Create a metric object}
\usage{
metric(
  metric,
  label = NA,
  colors = "grey60",
  point_size = 0.5,
  alpha = 1,
  type = "points"
)
}
\arguments{
\item{metric}{Metric name, should match a column in the input data frame
(i.e. the output of the \code{\link{load_genome_metrics}} function).}

\item{label}{Metric label, will be used in plot legends. If NA, the metric
name will be used (default: NA).}

\item{colors}{Colors, either be a string (e.g. "grey20") or a vector of
strings (e.g. c("blue", "yellow") (default: "grey60").}

\item{point_size}{Point size for plots of type "points", a float
(default: 0.5).}

\item{alpha}{Alpha value, a float (default: 1).}

\item{type}{Plot type for circos and region tracks, either "points" or
"ribbon"; not used in manhattan plots (default: "points").}
}
\value{
A named list with the value of each metric property
}
\description{
Generate an object storing all properties for a metric
}
\examples{
fst <- metric("Fst", color = "grey70", point_size = 0.75)
}
