% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{multi_metrics_track}
\alias{multi_metrics_track}
\title{Create a track object with multiple metrics}
\usage{
multi_metrics_track(
  metrics,
  metric_labels = NA,
  colors = NA,
  point_size = NA,
  alpha = NA,
  type = NA,
  label = NA,
  bg_colors = NA,
  ylim = NA,
  major_lines_x = NA,
  major_lines_y = NA,
  legend_position = NA
)
}
\arguments{
\item{metrics}{Vector of metrics to represent in the track, should correspond
to column names in the input data frame (i.e. the output of the
\code{\link{load_genome_metrics}} function).}

\item{metric_labels}{Vector of metric labels, will be used in plot legends.
If NA, metric names will be used (default: NA).}

\item{colors}{Either a single color (e.g. "grey20") which will be
applied to all values, a vector of colors (e.g. c("blue", "yellow")) which
will be alternatingly applied to linkage groups / chromosomes, or a function
returning a color for a given value.}

\item{point_size}{Point size for plots of type "points", can be a single
float value (e.g. 0.5) and will be applied to all metrics, or a vector of
floats of length equal to the number of metrics (e.g. c(0.5, 1, 0.75) for
three metrics).}

\item{alpha}{Alpha value, can be a single float value (e.g. 0.5) and will be
applied to all metrics, or a vector of floats of length equal to the number
of metrics (e.g. c(0.5, 0.8) for two metrics).}

\item{type}{Plot type for circos and region tracks, either "points" or
"ribbon"; can be a single value and will be applied to all metrics, or a
vector of strings of length equal to the number of metrics
(e.g. c("points", "points", "ribbon") for three metrics).}

\item{label}{Label to use for the track, will be displayed on the y axis.}

\item{bg_colors}{Background colors for circos and manhattan plots, not used
in region plots. Values can be a string (e.g. "white") and will be applied
to all sectors / chromosomes, or a vector of strings
(e.g. c("white", "grey80")) to assign alternating values to
sectors / chromosomes.}

\item{ylim}{Y-axis limits, a vector with values c(min, max) or NA to infer
y limits from the data.}

\item{major_lines_x}{If TRUE, reference lines will be plotted for the x axis,
equivalent to panel.grid.major.x in \code{\link{ggplot2::theme}}.}

\item{major_lines_y}{If TRUE, reference lines will be plotted for the y axis,
equivalent to panel.grid.major.y in \code{\link{ggplot2::theme}}.}

\item{legend_position}{Position of the legend, directly passed to
"legend.position" in \code{\link{ggplot2::theme}} in region plots, not used
in manhattan and in circos plots.}
}
\value{
A named list with the value of each track property
}
\description{
Simplified interface to generate a track representing multiple
metric.s All parameters have default values as NA, meaning their value
will be taken from default values specified in high-level functions.
}
\examples{
# Create a track object labeled "SNPs" for the metrics "Snp_females" in blue
# and "Snp_males" in red plotted as ribbons with alpha value of 0.5.
snps_track <- track(c("Snp_females", "Snp_males"),
                    label = "SNPs",
                    colors = c("firebrick2", "dodgerblue3"),
                    type = "ribbon",
                    alpha = 0.5)
}
