% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_region}
\alias{parse_region}
\title{Parse region}
\usage{
parse_region(region, contig_lengths)
}
\arguments{
\item{region}{String defining the region, either "Contig" or
"Contig:start-end"}

\item{contig_lengths}{Named vector with contig identifers as names and
contig lengths as values
(e.g. output of the \code{\link{load_genome_input}} function)}
}
\value{
A list of three elements: contig name, region start, and region end
}
\description{
Parse a region string with format "Contig:start-end"
}
\examples{
data <- load_genome_input("psass_window.tsv")
region_info <- parse_region("Chr01:150000-250000", data$lengths)
}
