% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{plot_circos}
\alias{plot_circos}
\title{Plot genome metrics in a circular layout}
\usage{
plot_circos(
  input_file_path,
  tracks,
  chromosomes_file_path = NULL,
  detect_chromosomes = TRUE,
  unplaced_label = "Unplaced",
  highlight = NULL,
  output_file = NULL,
  width = 2400,
  height = 2400,
  res = 120,
  default_color = "grey20",
  default_point_size = 0.25,
  default_ylim = NULL,
  default_bg_color = "white",
  default_bg_highlight_color = "grey80",
  sector_titles_expand = NULL
)
}
\arguments{
\item{input_file_path}{Path to a genome metrics input file (e.g. result of
PSASS or RADSex).
Format: Contig | Position | Length | <Metric1> | <Metric2> ... with
Contig = contig identifier, Position = position on the contig, Length =
length of the contig, <MetricN> = value of metric N (e.g. Fst) at the
corresponding position on the corresponding contig.}

\item{tracks}{List of track data objects for each track to include in the
plot. Track data objects for circos plots are generated with the
\code{\link{circos_track}} function. Tracks can represent one or mutiple
metrics from the genome metrics input file.}

\item{chromosomes_file_path}{Path to the chromosome names file
(i.e. tab-separated file without header and with columns
<Contig ID> | <Chromosome name>). If NULL, all contigs will be considered
unplaced except if detect_chromosomes is set to TRUE, in which case
chromosomes will be detected automatically from contig identifiers
(default: NULL)}

\item{detect_chromosomes}{If TRUE, will consider contigs starting with
"LG", "CHR", or "NC" as chromosomes if no chromosomes were specified
(default: TRUE)}

\item{unplaced_label}{Label for unplaced contigs superscaffold
(default: "Unplaced")}

\item{highlight}{Vector containing the names or identifiers of contigs or
chromosomes to highlight in the circos plot (default NULL)}

\item{output_file}{Path to an output file for the generated circos plot,
or NULL to plot in the current R device (default: NULL)}

\item{width}{Plot width when plotting to an output file, in pixel
(default: 2400)}

\item{height}{Plot height when plotting to an output file, in pixel
(default: 2400)}

\item{res}{Image resolution when plotting to an output file, in ppi
(default: 120)}

\item{default_color}{Default color for a track when not specified
in track data (default: "grey20")}

\item{default_point_size}{Default point size for a track when not specified
in track data (default: 0.25)}

\item{default_ylim}{Default y-axis limits for a track when not specified
in track data (default: NULL, i.e. infer from data)}

\item{default_bg_color}{Default background color for standard sectors
in the track when not specified in track data (default: "white")}

\item{default_bg_highlight_color}{Default background color for highlighted
sectors in the track when not specified in track data (default: "grey80")}

\item{sector_titles_expand}{Manually set the space between sector titles
and x-axis as a multiple of ymax (default: NULL)}
}
\description{
Generate a circos plot with multiple tracks from a
genome metrics file
}
\examples{
plot_circos("data/psass_window.tsv",
            tracks = list(track("Fst",
                                label = expression("F"["ST"])),
                          track(c("Snps_females", "Snps_males"),
                                label = "Pool-specific SNPs",
                                color = c("firebrick2", "dodgerblue3")),
                          track(c("Depth_ratio"),
                                label = "Depth ratio",
                                color = "grey50")),
            output_file = "circos.png")
}
