% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{plot_circos_track}
\alias{plot_circos_track}
\title{Plot a circos track}
\usage{
plot_circos_track(
  data,
  track,
  top.track = FALSE,
  sector_titles_expand = 1.3,
  first.sector = NULL
)
}
\arguments{
\item{data}{Input data generated with the \code{\link{create_region_track_data}} function)}

\item{track}{Track object storing properties for the current track, generated with the \code{\link{circos_track}} function}

\item{top.track}{If TRUE, x-axis labels and sector names will be added to the track}

\item{sector_titles_expand}{Manually set the space between sector titles and x-axis as a multiple of ymax (default: 1.3)}

\item{first.sector}{Name of the first sector to draw y-axis on (default: NULL)}
}
\description{
Plot a single track for a circos plot
}
\examples{
genomic_data <- load_genome_input("psass_window.tsv")
fst_track <- circos_track("Fst")
circos_data <- create_circos_track_data(genomic_data, fst_track)

plot_track_region(region_data, fst_track, top.track=TRUE)

}
