% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_track_manhattan}
\alias{plot_track_manhattan}
\title{Plot a manhattan track}
\usage{
plot_track_manhattan(
  data,
  backgrounds,
  track,
  bottom.track = FALSE,
  x.title = "Chromosomes",
  show.chromosomes.names = TRUE,
  chromosomes.as.numbers = FALSE
)
}
\arguments{
\item{data}{Backgrounds data generated with the \code{\link{assign_manhattan_background_colors}} function)}

\item{track}{Track object storing properties for the current track, generated with the \code{\link{manhattan_track}} function}

\item{bottom.track}{If TRUE, x-axis labels and title will be added to the plot}

\item{x.title}{Title of the x-axis}

\item{show.chromosomes.names}{If TRUE, display chromosome names on the x axis}

\item{chromosomes.as.numbers}{If TRUE, display chromosome numbers instead of names for readability}
}
\value{
A ggplot object for the manhattan plot
}
\description{
Plot a single track for a manhattan plot
}
\examples{
genomic_data <- load_genome_input("psass_window.tsv")
fst_track <- manhattan_track("Fst")
manhattan_data <- create_manhattan_track_data(genomic_data, fst_track)

fst_plot <- plot_track_manhattan(manhattan_data, fst_track, bottom.track=TRUE)

}
