% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{plot_track_region}
\alias{plot_track_region}
\title{Plot a region track}
\usage{
plot_track_region(data, region.info, track, bottom.track = FALSE)
}
\arguments{
\item{data}{Input data generated with the \code{\link{create_region_track_data}} function)}

\item{region.info}{Information on the region to plot, output of the \code{\link{parse_region}} function}

\item{track}{Track object storing properties for the current track, generated with the \code{\link{region_track}} function}

\item{bottom.track}{If TRUE, x-axis labels and title will be added to the plot}
}
\value{
A ggplot object for the plot
}
\description{
Plot a single track for a genomic region
}
\examples{
genomic_data <- load_genome_input("psass_window.tsv")
region_info <- parse_region("Chr01:0-1500000")
fst_track <- region_track("Fst")
region_data <- create_region_track_data(genomic_data, region_info, fst_track)

fst_plot <- plot_track_region(region_data, region_info, fst_track, bottom.track=TRUE)

}
