% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsex_depth.R
\name{radsex_depth}
\alias{radsex_depth}
\title{Plot the results of radsex "depth" analysis}
\usage{
radsex_depth(
  input_file,
  metric = "Markers",
  type = "boxplot",
  groups = NA,
  group_labels = NA,
  group_colors = c("dodgerblue3", "firebrick2"),
  alpha = 0.8,
  output_file = NA,
  width = 8,
  height = 8,
  res = 300,
  title = NA,
  ylab = NA,
  major_lines_y = TRUE,
  pretty_y_scale = TRUE,
  boxplot_jitter = TRUE,
  barplot_order_by = "Metric"
)
}
\arguments{
\item{input_file}{Path to a table of marker depth statistics per individual
obtained with radsex "depth".}

\item{metric}{Metric to display on the y axis, one of the columns of the
output of radsex "depth": "Reads", "Markers", "Min_depth", "Max_depth",
"Median_depth", "Average_depth" (default: "Markers").}

\item{type}{Plot type, either "boxplot" or "barplot" (default: "boxplot").}

\item{groups}{Vector specifying which groups to include in the plot, ordered.
Should match group names in the input data. If NA, groups will be infered
from the input data (default: NA).}

\item{group_labels}{Vector specifying the label to associate to each group
in the plot axis / legend, or NA to use the group names (default: NA).}

\item{group_colors}{Vector specyfing the color to use for each group in the
plot (default: c("dodgerblue3", "firebrick2")).}

\item{alpha}{Alpha value for plot elements, a float (default: 0.8).}

\item{output_file}{Path to an output file for the generated boxplot,
or NA to plot in the current R device (default: NA).}

\item{width}{Plot width when plotting to an output file, in inches
(default: 8).}

\item{height}{Plot height when plotting to an output file, in inches
(default: 8).}

\item{res}{Image resolution when plotting to an output file, in dpi
(default: 300).}

\item{title}{Plot title, or NA for no title (default: NA).}

\item{ylab}{Title of the y axis, or NA to use the metric name (default: NA).}

\item{major_lines_y}{If TRUE, reference lines will be plotted for the y axis,
equivalent to panel.grid.major.y in \code{\link{ggplot2::theme}}
(default: TRUE).}

\item{pretty_y_scale}{If TRUE, generate pretty labels for the y-axis scale
and add the unit (M. or K.) to the axis title (default: TRUE).}

\item{boxplot_jitter}{If TRUE, superpose the distribution of values as points
in the boxplot (default: TRUE).}

\item{barplot_order_by}{Value to use to order bars in barplots, either
"Group", "Individual", or "Metric" (default: "Metric").}
}
\value{
A ggplot object for the plot
}
\description{
Plot any metric computed with radsex "depth", either as a
boxplot showing the distribution of the metric for each group or as a barplot
showing the metric value for each individual.
}
\examples{
reads_plot <- radsex_depth("depth.tsv",
                           groups = c("M", "F"),
                           group_labels = c("Males", "Females"),
                           group_colors = c("green", "purple"),
                           metric = "Reads",
                           ylab = "Number of reads")

depth_plot <- radsex_depth("depth.tsv",
                           by_group = FALSE,
                           group_colors = c("dodgerblue3", "red2"),
                           metric = "Median_depth",
                           ylab = "Number of reads",
                           output_file = "median_depth.png")
}
