% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsex_distrib.R
\name{radsex_distrib}
\alias{radsex_distrib}
\title{Plot the results of radsex "distrib" analysis}
\usage{
radsex_distrib(
  input_file,
  groups = NA,
  group_labels = NA,
  output_file = NA,
  width = 8,
  height = 7,
  res = 300,
  autoscale = TRUE,
  title = NA,
  show_significance = TRUE,
  significance_color = "red3",
  significance_thickness = 0.5,
  bins = c(0, 1, 5, 25, 100, 1000),
  colors = c("white", "navyblue")
)
}
\arguments{
\item{input_file}{Path to a table of distribution of markers between groups
obtained with radsex "distrib".}

\item{groups}{Vector of length 2 with the groups to plot on the x and y axes,
should match column headers in the input data. If NA, groups will be infered
from the input data (default: NA).}

\item{group_labels}{Vector of length 2 with the labels to associate to each
group in the plot axis titles, or NA to use the group names (default: NA).}

\item{output_file}{Path to an output file for the generated distrib plot,
or NA to plot in the current R device (default: NA).}

\item{width}{Plot width when plotting to an output file, in inches
(default: 8).}

\item{height}{Plot height when plotting to an output file, in inches
(default: 7).}

\item{res}{Image resolution when plotting to an output file, in dpi
(default: 300).}

\item{autoscale}{If TRUE, the width and height of the output file will be
adjusted so that horizontal and vertical scales are the same. The resulting
plot will not be square if there are different number of individuals in each
group (default: TRUE).}

\item{title}{Plot title (default: NA, i.e. no title).}

\item{show_significance}{If TRUE, tiles for which association with group is
significant are highlighted using the color defined with the
"significance_color" parameter (default: TRUE).}

\item{significance_color}{Color of the border for tiles in which association
with group is significant (default: "red3").}

\item{significance_thickness}{Thickness of the border for tiles in which
association with group is significant (default: 0.5).}

\item{bins}{A vector of values to use as bins in the color palette
(default: c(0, 1, 5, 25, 100, 1000)).}

\item{colors}{A vector of two colors used to create the color palette
gradient (default: c("white", "navyblue")).}
}
\value{
A ggplot object for the tile plot
}
\description{
Plot a distribution of markers between two groups obtained with
radsex "distrib". In the resulting plot, the color of a tile at coordinates
(x, y) indicates the number of markers present in x individuals from group1
and y individuals from group2.
}
\examples{
distrib_plot <- radsex_distrib("distrib.tsv",
                               groups = c("M", "F"),
                               group_labels = c("Males", "Females"),
                               title = "Distribution of markers",
                               significance_color = "green",
                               bins = c(0, 10, 100, 1000),
                               colors = c("white", "red3"))
}
