% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsex_freq.R
\name{radsex_freq}
\alias{radsex_freq}
\title{Plot the results of radsex "freq" analysis}
\usage{
radsex_freq(
  input_file,
  color = "grey60",
  alpha = 0.8,
  output_file = NA,
  width = 10,
  height = 6,
  res = 300,
  title = NA,
  ylab = "Marker presence (density)",
  xlab = "Number of samples",
  major_lines_y = TRUE
)
}
\arguments{
\item{input_file}{Path to a table of marker presence in all individuals
obtained with radsex "freq".}

\item{color}{Plot main color (default: "grey60").}

\item{alpha}{Alpha value for plot elements, a float (default: 1).}

\item{output_file}{Path to an output file for the generated boxplot,
or NA to plot in the current R device (default: NA).}

\item{width}{Plot width when plotting to an output file, in inches
(default: 12).}

\item{height}{Plot height when plotting to an output file, in inches
(default: 4).}

\item{res}{Image resolution when plotting to an output file, in dpi
(default: 300).}

\item{title}{Plot title, or NA for no title (default: NA).}

\item{major_lines_y}{If TRUE, reference lines will be plotted for the y axis,
equivalent to panel.grid.major.y in \code{\link{ggplot2::theme}}
(default: TRUE).}

\item{type}{Plot type, either "boxplot" or "barplot" (default: "barplot").}

\item{pretty_y_scale}{If TRUE, generate pretty labels for the y-axis scale
and add the unit (M. or K.) to the axis title (default: TRUE).}

\item{boxplot_jitter}{If TRUE, superpose the distribution of values as points
in the boxplot (default: TRUE).}

\item{barplot_frequencies}{If TRUE, display values as frequency between 0
and 1 instead of absolute values (default: FALSE).}
}
\value{
A ggplot object for the plot
}
\description{
Plot any metric computed with radsex "freq", either as a single
boxplot showing the distribution of markers in all individuals or as a
barplot showing the metric value for each number of individuals.
}
\examples{
freq_plot <- radsex_freq("depth.tsv",
                         type = "boxplot",
                         color = "purple",
                         ylab = "Number of markers",
                         output_file = "markers_freq.png")
}
