% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsex_map.R
\name{radsex_map_circos}
\alias{radsex_map_circos}
\title{RADSex circos plot}
\usage{
radsex_map_circos(
  input_file,
  chromosomes_file = NA,
  detect_chromosomes = TRUE,
  unplaced_label = "U.",
  output_file = NA,
  width = 2200,
  height = 2200,
  res = 120,
  tracks = c("p", "bias"),
  bias_colors = c("firebrick1", "black", "dodgerblue2"),
  p_color = "grey40",
  point_size = 0.3,
  highlight = NA,
  highlight_bg_color = "grey80",
  sector_titles_expand = NA
)
}
\arguments{
\item{input_file}{Path to the result of radsex "map".
Format: Contig | Position | Length | Marker ID | Bias | P | Signif
Contig = contig identifier, Position = position on the contig, Length =
length of the contig, <MetricN> = value of metric N at the
corresponding position on the corresponding contig.}

\item{chromosomes_file}{Path to the chromosome names file
(i.e. tab-separated file without header and with columns
<Contig ID> | <Chromosome name>). If NA, all contigs will be considered
unplaced except if detect_chromosomes is set to TRUE, in which case
chromosomes will be detected automatically from contig identifiers
(default: NA).}

\item{detect_chromosomes}{If TRUE, will consider contigs starting with
"LG", "CHR", or "NC" as chromosomes if no chromosomes were specified
(default: TRUE).}

\item{unplaced_label}{Label for unplaced contigs superscaffold
(default: "U.").}

\item{output_file}{Path to an output file for the generated circos plot,
or NA to plot in the current R device (default: NA)}

\item{width}{Plot width when plotting to an output file, in pixel
(default: 2400).}

\item{height}{Plot height when plotting to an output file, in pixel
(default: 2400).}

\item{res}{Image resolution when plotting to an output file, in ppi
(default: 120).}

\item{tracks}{Vector of tracks to plot, in order from top to bottom; possible
tracks: "bias" for bias between groups, "p" for probability of association
with group (default: c("p", "bias")).}

\item{bias_colors}{A vector of length 3 defining colors for extreme values in
the bias track, i.e. -1, 0, and 1
(default: c("firebrick1", "black", "dodgerblue2")).}

\item{p_color}{Color of the track showing -log(p of association with group)
(default: "grey20").}

\item{point_size}{Point size, a float (default: 0.3).}

\item{highlight}{Vector containing the names or identifiers of contigs or
chromosomes to highlight in the circos plot (default NA).}

\item{highlight_bg_color}{Background color for highlighted sectors
(default: "grey80").}

\item{sector_titles_expand}{Manually set the space between sector titles
and x-axis as a multiple of ymax (default: NA).}
}
\description{
Generates a circular plot of radsex "map" results in which each
sector represents a linkage group / chromosome and the x-axis represents the
position on the linkage group. The y-axis on the first track shows the bias
of a marker between groups and the second track shows the probability of
association with group for a marker.
}
\examples{
radsex_map_circos("radsex_map.tsv",
                  chromosomes_file = "chromosomes.tsv")
}
