% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsex_map.R
\name{radsex_map_manhattan}
\alias{radsex_map_manhattan}
\title{RADSex manhattan plot}
\usage{
radsex_map_manhattan(
  input_file,
  chromosomes_file = NA,
  detect_chromosomes = TRUE,
  unplaced_label = "U.",
  output_file = NA,
  width = 12,
  height = 6,
  res = 300,
  colors = c("dodgerblue3", "darkgoldenrod2"),
  bg_colors = c("grey85", "white"),
  point_size = 0.5,
  x_title = NA,
  show_chromosomes_names = TRUE,
  chromosomes_as_numbers = FALSE,
  show_signif_line = TRUE,
  signif_line_color = "red"
)
}
\arguments{
\item{input_file}{Path to the result of radsex "map".
Format: Contig | Position | Length | Marker ID | Bias | P | Signif
Contig = contig identifier, Position = position on the contig, Length =
length of the contig, <MetricN> = value of metric N at the
corresponding position on the corresponding contig.}

\item{chromosomes_file}{Path to the chromosome names file
(i.e. tab-separated file without header and with columns
<Contig ID> | <Chromosome name>). If NA, all contigs will be considered
unplaced except if detect_chromosomes is set to TRUE, in which case
chromosomes will be detected automatically from contig identifiers
(default: NA).}

\item{detect_chromosomes}{If TRUE, will consider contigs starting with
"LG", "CHR", or "NC" as chromosomes if no chromosomes were specified
(default: TRUE).}

\item{unplaced_label}{Label for unplaced contigs superscaffold
(default: "U.").}

\item{output_file}{Path to an output file for the generated circos plot,
or NA to plot in the current R device (default: NA)}

\item{width}{Plot width when plotting to an output file, in inches
(default: 12).}

\item{height}{Plot height when plotting to an output file in inches,
(default: 6).}

\item{res}{Image resolution when plotting to an output file, in dpi
(default: 300).}

\item{colors}{A single color value for points that will be applied to all
chromosomes or a vector of alternating colors
(default: c("dodgerblue3", "darkgoldenrod2")).}

\item{bg_colors}{A single color value for background that will be applied to
all chromosomes or a vector of alternating colors
(default: c("grey85", "white")).}

\item{point_size}{Point size, a float (default: 1).}

\item{x_title}{Title of the x-axis (default: "Chromosomes").}

\item{show_chromosomes_names}{If TRUE, display chromosome names on the x axis
(default: TRUE).}

\item{chromosomes_as_numbers}{If TRUE, display chromosome numbers instead of
names for readability (default: FALSE).}

\item{show_signif_line}{If TRUE, show a horizontal line at the p-value
threshold for significance with Bonferroni correction (default: TRUE).}

\item{signif_line_color}{Color of the significance line (default: "red").}
}
\value{
A ggplot object for the plot.
}
\description{
Generates a Manhattan plot of radsex "map" results in which
-log10(probability of association with group) is plotted for all markers.
}
\examples{
radsex_map_manhattan("radsex_map.tsv",
                     output_file = "manhattan.svg",
                     colors = c("green", "purple"),
                     chromosomes_as_numbers = TRUE)
}
