% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsex_map.R
\name{radsex_map_region}
\alias{radsex_map_region}
\title{RADSex region plot}
\usage{
radsex_map_region(
  input_file,
  region,
  chromosomes_file = NA,
  detect_chromosomes = TRUE,
  output_file = NA,
  width = 12,
  track_height = 4,
  res = 300,
  tracks = c("p", "bias"),
  bias_colors = c("firebrick1", "black", "dodgerblue2"),
  p_color = "grey40",
  type = "points",
  point_size = 1,
  alpha = 1
)
}
\arguments{
\item{input_file}{Path to the result of radsex "map".
Format: Contig | Position | Length | Marker ID | Bias | P | Signif
Contig = contig identifier, Position = position on the contig, Length =
length of the contig, <MetricN> = value of metric N at the
corresponding position on the corresponding contig.}

\item{region}{Region to plot, defined with the syntax "Contig" or
"Contig:start-end"}

\item{chromosomes_file}{Path to the chromosome names file
(i.e. tab-separated file without header and with columns
<Contig ID> | <Chromosome name>). If NA, all contigs will be considered
unplaced except if detect_chromosomes is set to TRUE, in which case
chromosomes will be detected automatically from contig identifiers
(default: NA).}

\item{detect_chromosomes}{If TRUE, will consider contigs starting with
"LG", "CHR", or "NC" as chromosomes if no chromosomes were specified
(default: TRUE).}

\item{output_file}{Path to an output file for the generated region plot,
or NA to plot in the current R device (default: NA)}

\item{width}{Plot width when plotting to an output file, in inches
(default: 12).}

\item{track_height}{Height of a single track when plotting to an output file,
in inches (default: 4).}

\item{res}{Image resolution when plotting to an output file, in dpi
(default: 300).}

\item{tracks}{Vector of tracks to plot, in order from top to bottom; possible
tracks: "bias" for bias between groups, "p" for probability of association
with group (default: c("p", "bias")).}

\item{bias_colors}{A vector of length 3 defining colors for extreme values in
the bias track, i.e. -1, 0, and 1
(default: c("firebrick1", "black", "dodgerblue2")).}

\item{p_color}{Color of the track showing -log(p of association with group)
(default: "grey20").}

\item{point_size}{Point size, a float (default: 1).}

\item{alpha}{Alpha value, a float (default: 1).}
}
\value{
A ggplot object for the plot.
}
\description{
Generates a linear plot of radsex "map" results for a given
genomic region, where the x-axis represents the position on the contig,
the y-axis on the first track shows the probability of association with group
for a marker and the second track shows the bias of a marker between groups,
}
\examples{
region_plot <- radsex_map_region("radsex_map.tsv",
                                 region = "Contig0001:1500-3500",
                                 tracks = c("bias"))
}
