% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genome_metrics.R
\name{read_comments}
\alias{read_comments}
\title{Parse comment lines in an input file}
\usage{
read_comments(
  input_file,
  comment_char = "#",
  comment_sep = ";",
  comment_internal_sep = ":"
)
}
\arguments{
\item{input_file}{Path to an input file.}

\item{comment_char}{Character indicating a comment line (default: "#").}

\item{comment_sep}{Character separating two fields in a comment line
(default: ";").}

\item{comment_internal_sep}{Character separating property and value in a
field from a comment line (default: ":").}
}
\value{
A named list of property values, or an empty list if no property
could be found in the input file.
}
\description{
Extract and parse comment lines from an input file. Comment
lines are expected to follow the pattern:
<comment_char>property<comment_internal_sep>value<comment_sep>property...
With default values: #property1:value1;property2:value2...
}
\examples{
data <- load_table("subset.tsv")
}
