% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{region_track}
\alias{region_track}
\title{Create region track object}
\usage{
region_track(
  metrics,
  label = NULL,
  color = NULL,
  alpha = NULL,
  type = NULL,
  point.size = NULL,
  major.lines.y = NULL,
  major.lines.x = NULL,
  legend.position = NULL,
  ylim = NULL
)
}
\arguments{
\item{metrics}{Metrics included in the track. Metrics should correspond to column names in the data frame used as input
data in the plot (output of the \code{\link{load_genome_input}} function). Possible values: a string if the track includes
a single metric (e.g. "Fst"), or a vector if the track includes several metrics (e.g. c("Snps_females", "Snps_males"))}

\item{label}{Track label, NULL to set the label to the metric name for single-metric tracks. Label has to be specified for multi-metrics tracks (default: NULL)}

\item{color}{Track color. Values can be a string (e.g. "grey20") and will then be applied to all metrics, or a vector of size equal to the number of metrics
(e.g. c("red", "blue") for two metrics)}

\item{alpha}{Track alpha value. Values can be a float (e.g. 0.5) and will then be applied to all metrics, or a vector of size equal to the number of metrics
(e.g. c(1, 0.5, 0.75) for three metrics)}

\item{type}{Track plot type. Possible values: "ribbon" or "points". Values can be a string (e.g. "ribbon") and will then be applied to all metrics,
or a vector of size equal to the number of metrics (e.g. c("points", "ribbon") for two metrics)}

\item{point.size}{Point size for plots of type "points". Values can be a float (e.g. 0.5) and will then be applied to all metrics,
or a vector of size equal to the number of metrics (e.g. c(1, 1.5, 3) for three metrics)}

\item{major.lines.y}{Plot reference lines for the y axis, directly passed to "major.lines.y" in \code{\link{ggplot2::theme}} (single value)}

\item{major.lines.x}{Plot reference lines for the x axis, directly passed to "major.lines.x" in \code{\link{ggplot2::theme}} (single value)}

\item{legend.position}{Position of the legend for the track, directly passed to "legend.position" in \code{\link{ggplot2::theme}} (single value)}

\item{ylim}{Vector of y-axis limits for the track; if NULL, infer directly from data}
}
\value{
A named list with the value of each track property
}
\description{
Generate an object storing all properties for a region track
}
\examples{
# Single metric
track_data <- region_track("Fst", color = "grey70", type = "points", point.size = 0.75)

# Multiple metrics
track_data <- region_track(c("Snp_females", "Snp_males"), color = c("firebrick2", "dodgerblue3"), alpha = 0.5)

}
