% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Create a track object}
\usage{
track(
  metrics,
  label = NA,
  bg_colors = NA,
  ylim = NA,
  major_lines_x = NA,
  major_lines_y = NA,
  legend_position = NA,
  h_lines = NA
)
}
\arguments{
\item{metrics}{Metrics to represent in the track; metric names are given by
column names in the input data frame (i.e. the output of the
\code{\link{load_genome_metrics}} function). Possible values: a string if
the track represents a single metric (e.g. "Fst"), or a vector if the track
represents several metrics (e.g. c("Snps_females", "Snps_males")).}

\item{label}{Label to use for the track, will be displayed on the y axis. Can
be set to NA for single metric tracks to set the label to the name of the
metric, but label has to be specified for multi-metrics tracks.}

\item{bg_colors}{Background colors for circos and manhattan plots. Values can
be a string (e.g. "white") and will be applied to all sectors / chromosomes,
or a vector of colors (e.g. c("white", "grey80")) to assign
alternating values to sectors / chromosomes. Not used in region plots.}

\item{ylim}{Y-axis limits, a vector with values c(min, max).}

\item{major_lines_x}{If TRUE, reference lines will be plotted for the x axis.
The value for this parameter is directly passed to "major.lines.x" in
\code{\link{ggplot2::theme}} for region plots. Not used in manhattan plots.}

\item{major_lines_y}{If TRUE, reference lines will be plotted for the y axis.
The value for this parameter is directly passed to "major.lines.y" in
\code{\link{ggplot2::theme}} for manhattan and region plots.}

\item{legend_position}{Position of the legend, directly passed to
"legend.position" in \code{\link{ggplot2::theme}} for region plots. Not used
in manhattan and in circos plots.}

\item{h_lines}{List of objects defined with \code{\link{h_line}} adding
horizontal lines to the track (default: NA).}
}
\value{
A named list with the value of each track property
}
\description{
Generate an object storing all properties for a plot track
}
\examples{
# Single metric
track_data <- track("Fst", color = "grey70", point_size = 0.75)

# Multiple metrics
track_data <- track(c("Snp_females", "Snp_males"), label = "SNPs",
                    colors = c("firebrick2", "dodgerblue3"))
}
