% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{draw_circos}
\alias{draw_circos}
\title{Draw genome metrics in a circular layout}
\usage{
draw_circos(
  data,
  contig_lengths,
  tracks,
  highlight = NULL,
  output_file = NULL,
  width = 2400,
  height = 2400,
  res = 120,
  default_color = "grey20",
  default_point_size = 0.25,
  default_ylim = NULL,
  default_bg_color = "white",
  default_bg_highlight_color = "grey80",
  sector_titles_expand = NULL
)
}
\arguments{
\item{data}{Genome metrics data frame (e.g. result of PSASS or RADSex loaded
with the \code{\link{load_genome_input}} function).
Format: Contig | Position | Length | <Metric1> | <Metric2> ... with
Contig = contig identifier, Position = position on the contig, Length =
length of the contig, <MetricN> = value of metric N (e.g. Fst) at the
corresponding position on the corresponding contig.}

\item{contig_lengths}{Contig lengths data frame from the output of the
\code{\link{load_genome_input}} function. Format: <Contig ID> | <Length>}

\item{tracks}{List of track data objects for each track to include in the
plot. Track data objects for circos plots are generated with the
\code{\link{circos_track}} function. Tracks can represent one or mutiple
metrics from the genome metrics input file.}

\item{highlight}{Vector containing the names or identifiers of contigs or
chromosomes to highlight in the circos plot (default NULL)}

\item{output_file}{Path to an output file for the generated circos plot,
or NULL to plot in the current R device (default: NULL)}

\item{width}{Plot width when plotting to an output file, in pixel
(default: 2400)}

\item{height}{Plot height when plotting to an output file, in pixel
(default: 2400)}

\item{res}{Image resolution when plotting to an output file, in ppi
(default: 120)}

\item{default_color}{Default color for a track when not specified
in track data (default: "grey20")}

\item{default_point_size}{Default point size for a track when not specified
in track data (default: 0.25)}

\item{default_ylim}{Default y-axis limits for a track when not specified
in track data (default: NULL, i.e. infer from data)}

\item{default_bg_color}{Default background color for standard sectors
in the track when not specified in track data (default: "white")}

\item{default_bg_highlight_color}{Default background color for highlighted
sectors in the track when not specified in track data (default: "grey80")}

\item{sector_titles_expand}{Manually set the space between sector titles
and x-axis as a multiple of ymax (default: NULL)}
}
\description{
Generate a circos plot with multiple tracks from a
genome metrics data frame and a contig lengths vector
}
\examples{
metrics <- load_genome_metrics("psass_window.tsv")
plot_circos(metrics$data,
            metrics$lengths,
            tracks = list(track("Fst",
                                label = expression("F"["ST"])),
                          track(c("Snps_females", "Snps_males"),
                                label = "Pool-specific SNPs",
                                color = c("firebrick2", "dodgerblue3")),
                          track(c("Depth_ratio"),
                                label = "Depth ratio",
                                color = "grey50")),
            output_file = "circos.png")
}
