% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genome_metrics.R
\name{load_chromosome_names}
\alias{load_chromosome_names}
\title{Load chromosome names}
\usage{
load_chromosome_names(input_file)
}
\arguments{
\item{input_file}{Path to the chromosome names file (i.e. tab-separated
file without header and with columns <Contig ID> | <Chromosome name>).}
}
\value{
A named vector with contig identifiers as names and chromosome names
as values, or NA if the input file path is NA.
}
\description{
Loads chromosome names from a tabulated file
}
\examples{
chromosomes <- load_chromosome_names("chromosomes_names.tsv")
}
