% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genome_metrics.R
\name{load_genome_metrics}
\alias{load_genome_metrics}
\title{Load a genome metrics data file}
\usage{
load_genome_metrics(
  input_file,
  chromosomes = NA,
  detect_chromosomes = TRUE,
  unplaced_label = "Unplaced",
  comment_char = "#",
  comment_sep = ";",
  comment_internal_sep = ":"
)
}
\arguments{
\item{input_file}{Path to a genome metrics file (format described above)}

\item{chromosomes}{Named vector of chromosome names from
\code{\link{load_chromosome_names}} (default: NA)}

\item{detect_chromosomes}{If TRUE, will consider contigs starting with
"LG", "CHR", or "NC" as chromosomes if no chromosomes were specified
(default: TRUE)}

\item{unplaced_label}{Label for unplaced contigs superscaffold
(default: "Unplaced")}

\item{comment_char}{Character indicating a comment line (default: "#").}

\item{comment_sep}{Character separating two fields in a comment line
(default: ";").}

\item{comment_internal_sep}{Character separating property and value in a
field from a comment line (default: ":").}
}
\value{
A list with three elements: "data" = parsed genome metrics data,
"lengths" = lengths of contigs included in the plot, "properties" = list
of properties from the comment lines.
}
\description{
Loads a file containing metric values along a genome.
Format: Contig | Position | Length | <Metric1> | <Metric2> ... with
Contig = contig identifier, Position = position on the contig, Length =
length of the contig, <MetricN> = value of metric N (e.g. Fst) at the
corresponding position on the corresponding contig.
}
\examples{
chromosomes <- load_chromosome_names("chromosomes_names.tsv")
genome_data <- load_genome_metrics("psass_window.tsv",
                                   chromosomes = chromosomes)
}
