% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radsex_data.R
\name{load_table}
\alias{load_table}
\title{Load a standard table file}
\usage{
load_table(
  input_file,
  get_properties = TRUE,
  comment_char = "#",
  comment_sep = ";",
  comment_internal_sep = ":"
)
}
\arguments{
\item{input_file}{Path to a tabulated file}
}
\value{
A data frame containing the tabulated file data.
}
\description{
Loads a tabulated file with readr
}
\examples{
data <- load_table("subset.tsv")
}
