% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{plot_region}
\alias{plot_region}
\title{Plot region}
\usage{
plot_region(
  input.file.path,
  region,
  tracks,
  chromosomes.file.path = NULL,
  detect.chromosomes = TRUE,
  output.file = NULL,
  width = 12,
  track.height = 4,
  res = 300,
  default.color = "grey20",
  default.alpha = 1,
  default.type = "ribbon",
  default.point.size = 0.5,
  default.ylim = NULL,
  default.major.lines.y = TRUE,
  default.major.lines.x = FALSE,
  default.legend.position = "none"
)
}
\arguments{
\item{input.file.path}{Path to a genomic data input file (e.g. result of PSASS or RADSex)}

\item{region}{Region to plot, with syntax "Contig" or "Contig:start-end"}

\item{tracks}{List of tracks to plot. Tracks can be generated with the \code{\link{region_track}} function}

\item{chromosomes.file.path}{Path to a tabulated file specifying chromosome names (default: NULL)}

\item{detect.chromosomes}{If TRUE, will consider contigs starting with LG, CH, or NC as chromosomes if no chromosomes were specified (default: TRUE)}

\item{output.file}{Path to an output file for the generated region plot, or NULL to plot in the current R device (default: NULL)}

\item{width}{Plot width when plotting to an output file, in inches (default: 12)}

\item{track.height}{Height of a single track when plotting to an output file, in inches (default: 4)}

\item{res}{Image resolution when plotting to an output file, in dpi (default: 300)}

\item{default.color}{Default color for a track when not specified in track data (default: "grey20")}

\item{default.alpha}{Default alpha value for a track when not specified in track data (default: 1)}

\item{default.type}{Default plot type for a track when not specified in track data (default: "ribbon")}

\item{default.point.size}{Default point size for a track of type "points" when not specified in track data (default: 0.5)}

\item{default.ylim}{Default y-axis limits for a track when not specified in track data (default: NULL, i.e. infer from data)}

\item{default.major.lines.y}{If TRUE, reference lines will be plotted for the y axis if not specified in track data (default: TRUE)}

\item{default.major.lines.x}{If TRUE, reference lines will be plotted for the x axis if not specified in track data (default: FALSE)}

\item{default.legend.position}{Default legend position when not specified in track data (default: "none")}
}
\value{
Combined plot data (ggplot object)
}
\description{
Generate a linear plot with multiple tracks for a specified genomic region from a genomic data file
}
\examples{

# Plotting an FST track and a combined pool-specific SNPs track
# for an entire chromosome to the default R device

plot_region("data/psass_window.tsv", "Chr01",
            tracks = list(region_track("Fst", label = expression("F"["ST"])),
                          region_track(c("Snps_females", "Snps_males"), label = "Pool-specific SNPs", color = c("firebrick2", "dodgerblue3"), alpha=0.6)),
            chromosomes.file.path = "data/chromosomes.tsv")

}
